FOIA-147: Load the most recent report revision.
Sending a report status back from "Published" to "Back with Agency"
creates a new revision of the node, but does not set that revision as
the default.  When validating uploaded report xml files, a report that
had been sent from "Published" to "Back with agency" was not allowed to
be processed, as it was still considered locked.  This is because the
default revision was being loaded (published), not the latest (back with
agency). This change ensures that the latest revision is checked during
this validation function.