FOIA-359: Cast string fields when numeric.
Some numeric values are stored in the database as string fields so that
the special values 'N/A' and '<1' can be used in those fields.  In order
to properly compare values during api requests, check if the value in
the database is numeric and cast it to an integer if it is being
compared to a numeric value.